backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1
sqlparse==0.5.3
